+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - YYYY-MM-DD
+### Added
+- Enumeration `components::table::Size::Weight(f64)`
+  to support distributing
+  the console width
+  not covered by `components::table::Size::Cells(usize)`
+  into weighted sections
 ## [0.7.0] - 2021-12-05
 ### Added
 - Table component